+++ inst/examples/02_sidebarPage/app.R
@@ -1,4 +1,3 @@
-require(shiny)
 require(saiUI)
 
 ui <- saiPage(